feat: Qwen integration improvements - fix duplicate output, add models, thinking chain support
Key changes:
- Fix Qwen SSE cumulative content causing duplicate output (add delta detection in sse_scanner.go)
- Add Qwen3-Max-Thinking, qwen3.6-plus, qwen3.6-plus-2026-04-02 model support
- Clean up qwen model list to only include web API supported models
- Add thinking chain framework for future Qwen deep think mode
- Update frontend API tester with new model options
- Various bug fixes and code cleanup

+++ internal/deepseek/pow.go
@@ -24,7 +24,7 @@ func (p *PowSolver) Compute(ctx context.Context, challenge map[string]any) (int6
 	expireAt := toInt64(challenge["expire_at"], 1680000000)
 	difficulty := toInt64FromFloat(challenge["difficulty"], 144000)
 
-	return pow.SolvePow(ctx, challengeStr, salt, expireAt, float64(difficulty))
+	return pow.SolvePow(ctx, challengeStr, salt, expireAt, difficulty)
 }
 
 func BuildPowHeader(challenge map[string]any, answer int64) (string, error) {
